build is working, load wdl does not
