+++ Makefile
@@ -2,7 +2,7 @@
 
 CC = gcc
 AR = ar
-CFLAGS = -MMD -fPIC -Wall -Wextra -O2 -g -I./include/ -I./src/common/include -I./src/game-state/include -I./src/ui/include -I./src/cli/include -I./src/action_management/include -I./src/checkpointing/include -I./src/game-state/src
+CFLAGS = -MMD -fPIC -Wall -Wextra -O0 -g -I./include/ -I./src/common/include -I./src/game-state/include -I./src/ui/include -I./src/cli/include -I./src/action_management/include -I./src/checkpointing/include -I./src/game-state/src
 RM = rm -f
 LDLIBS = -lyaml -lncursesw -lreadline -lprotobuf-c
 
@@ -23,7 +23,7 @@ all: $(BIN)
 #  - Makefile: src/ui/Makefile
 #  - Library: src/ui/ui.a
 
-COMPONENTS = libobj common ui cli checkpointing game-state action_management wdl
+COMPONENTS = common ui cli checkpointing game-state action_management wdl libobj
 
 LIBS = $(foreach comp,$(COMPONENTS),src/$(comp)/$(comp).a)
