+++ src/action_management/Makefile
@@ -5,7 +5,7 @@
 CC = gcc
 AR = ar
 SRCS = src/actionmanagement.c src/get_actions.c 
-CFLAGS = -MMD -Wall -Wextra -O2 -g -I./include/ -I../common/include/ -I../game-state/include/ -I../cli/include/ -I../ui/include/ -I../checkpointing/include -I../action-management/include -I../game-state/src
+CFLAGS = -MMD -Wall -Wextra -O0 -g -I./include/ -I../common/include/ -I../game-state/include/ -I../cli/include/ -I../ui/include/ -I../checkpointing/include -I../action-management/include -I../game-state/src
 OBJS = $(SRCS:.c=.o)
 LIB = action_management.a
 RM = rm -f
