+++ src/checkpointing/tests/Makefile
@@ -1,7 +1,7 @@
 # Makefile based on template at https://gist.github.com/xuhdev/1873316
 
 CC = gcc 
-CFLAGS = -fPIC -Wall -Wextra -O2 -g -I../include/ -I../../game-state/include -I../../game-state/src -I../../common/include -I../../action_management/include/
+CFLAGS = -fPIC -Wall -Wextra -O0 -g -I../include/ -I../../game-state/include -I../../game-state/src -I../../common/include -I../../action_management/include/
 RM = rm -f
 BIN = test-checkpointing
 LDLIBS = -lprotobuf-c -lcriterion
