+++ src/common/Makefile
@@ -2,7 +2,7 @@
 
 CC = gcc
 AR = ar
-CFLAGS = -MMD -fPIC -Wall -Wextra -O2 -g -I./include/ -I../game-state/include/ -I../ui/include/ -I../cli/include/ -I../action_management/include/ -I../checkpointing/include/ -I../game-state/src -I../wdl/include -I../libobj/include
+CFLAGS = -MMD -fPIC -Wall -Wextra -O0 -g -I./include/ -I../game-state/include/ -I../ui/include/ -I../cli/include/ -I../action_management/include/ -I../checkpointing/include/ -I../game-state/src -I../wdl/include -I../libobj/include
 RM = rm -f
 LIB = common.a
