+++ src/game-state/Makefile
@@ -1,6 +1,6 @@
 CC = gcc
 AR = ar
-CFLAGS = -fPIC -Wall -Wextra -O2 -g -I./include/ -I../common/include/ -I./src/ -I../action_management/include -I../action_management/src -I../ui/include -I../cli/include -I../wdl/include
+CFLAGS = -fPIC -Wall -Wextra -O0 -g -I./include/ -I../common/include/ -I./src/ -I../action_management/include -I../action_management/src -I../ui/include -I../cli/include -I../wdl/include
 DEPFLAGS = -MT $@ -MMD -MP -MF $(DEPDIR)/$*.d
 RM = rm -f
 SRCDIR = src
