+++ src/ui/Makefile
@@ -2,7 +2,7 @@
 
 CC = gcc
 AR = ar
-CFLAGS = -MMD -fPIC -Wall -Wextra -O2 -g -I./include/ -I../common/include/ -I../game-state/include/ -I../cli/include/ -I../action_management/include/ -I../checkpointing/include -I../game-state/src
+CFLAGS = -MMD -fPIC -Wall -Wextra -O0 -g -I./include/ -I../common/include/ -I../game-state/include/ -I../cli/include/ -I../action_management/include/ -I../checkpointing/include -I../game-state/src
 RM = rm -f
 LIB = ui.a
 LDLIBS = -lncursesw -lreadline
