+++ src/ui/tests/Makefile
@@ -1,5 +1,5 @@
 CC = gcc
-CFLAGS = -MMD -fPIC -Wall -Wextra -O2 -g -I../include/ -I../../common/include/ -I../../game-state/include/ -I../../cli/include/ -I../../action_management/include/
+CFLAGS = -MMD -fPIC -Wall -Wextra -O0 -g -I../include/ -I../../common/include/ -I../../game-state/include/ -I../../cli/include/ -I../../action_management/include/
 UI = -L../. -Wl,-rpath,../.
 COMMON = -L../../common/ -Wl,-rpath,../../common/ -l:common.a
 CLI = -L../../cli/ -Wl,-rpath,../../cli/ -l:cli.a
