+++ src/wdl/Makefile
@@ -2,11 +2,11 @@
 
 CC = gcc
 AR = ar
-CFLAGS = -MMD -fPIC -Wall -Wextra -O2 -g -I./include/ -I../libobj/include/ -I../game-state/include/ -I../action_management/include/ -I../common/include/ -I../ui/include -I../cli/include -I../checkpointing/include -I../game-state/src
+CFLAGS = -MMD -fPIC -Wall -Wextra -O0 -g -I./include/ -I../libobj/include/ -I../game-state/include/ -I../action_management/include/ -I../common/include/ -I../ui/include -I../cli/include -I../checkpointing/include -I../game-state/src
 RM = rm -f
 LIB = wdl.a
-LDFLAGS = -L../libobj -L../action_management
-LDLIBS = ../libobj/libobj.a ../action_management/action_management.a -lyaml
+LDFLAGS = -L../libobj -L../action_management -L../game-state
+LDLIBS = ../libobj/libobj.a ../action_management/action_management.a ../game-state/game-state.a -lyaml
 
 .PHONY: clean clean-lib clean-bin
 
@@ -14,7 +14,7 @@ all: $(LIB)
 
 # Library
 
-SRCS = src/parse.c src/validate.c src/load_room.c src/load_item.c src/load_game.c src/load_wdl.c
+SRCS = src/parse.c src/validate.c src/load_room.c src/load_item.c src/load_game.c
 OBJS = $(SRCS:.c=.o)
 
 -include $(SRCS:.c=.d)
