+++ src/wdl/include/load_item.h
@@ -11,6 +11,9 @@
 #include "parse.h"
 #include "game.h"
 #include "load_room.h"
+#include "wdl_common.h"
+#include "item.h"
+#include "game_action.h"
 
 #define FILE_PATH "../examples/sample_yaml/simplified_wdl.yaml"
