+++ test.yml
@@ -1,0 +1,52 @@
+#
+# Sample proof-of-concept yaml file
+#
+
+GAME:
+ - start: "room A"
+   intro: "This is the intro"
+
+ROOMS:
+ - id: "room A"
+   short_desc: "This is room A"
+   long_desc: "This is room A long"
+   connections:
+    - to: "room B"
+      direction: "north"
+
+ - id: "room B"
+   short_desc: "This is room B"
+   long_desc: "This is room B long"
+   connections:
+    - to: "room C"
+      direction: "north"
+
+
+ - id: "room C"
+   short_desc: "This is room C"
+   long_desc: "This is room C long"
+   connections:
+    - to: "room A"
+      direction: "north"
+
+ITEMS:
+ - id: "chair"
+   short_desc: "This is a chair"
+   long_desc: "This is a chair long"
+   in: "room A"
+   state: "still"
+   value: "no"
+
+ - id: "table"
+   short_desc: "This is a table"
+   long_desc: "This is a table long"
+   in: "room B"
+   state: "solid"
+   value: "no"
+
+ - id: "candle"
+   short_desc: "This is a candle"
+   long_desc: "This is a candle long"
+   in: "room C"
+   state: "hot"
+   value: "no"
